Rename FindCommand to FindCommandTest
