+++ src/test/java/seedu/organizer/logic/commands/FindNameCommandTestTest.java
@@ -15,28 +15,27 @@ import org.junit.Test;
 import seedu.organizer.logic.CommandHistory;
 import seedu.organizer.logic.UndoRedoStack;
 import seedu.organizer.logic.commands.exceptions.CommandException;
-import seedu.organizer.model.task.DescriptionContainsKeywordsPredicate;
+import seedu.organizer.model.task.NameContainsKeywordsPredicate;
 
-//@@author guekling
 /**
- * Contains integration tests (interaction with the Model) for {@code FindDescriptionCommand}.
+ * Contains integration tests (interaction with the Model) for {@code FindNameCommand}.
  */
-public class FindDescriptionCommandTest extends FindCommand<FindDescriptionCommand> {
+public class FindNameCommandTestTest extends FindCommandTest<FindNameCommand> {
     @Test
     public void equals() {
-        DescriptionContainsKeywordsPredicate firstPredicate =
-                new DescriptionContainsKeywordsPredicate(Collections.singletonList("cs2101"));
-        DescriptionContainsKeywordsPredicate secondPredicate =
-                new DescriptionContainsKeywordsPredicate(Collections.singletonList("CS2010"));
+        NameContainsKeywordsPredicate firstPredicate =
+                new NameContainsKeywordsPredicate(Collections.singletonList("first"));
+        NameContainsKeywordsPredicate secondPredicate =
+                new NameContainsKeywordsPredicate(Collections.singletonList("second"));
 
-        FindDescriptionCommand findFirstCommand = new FindDescriptionCommand(firstPredicate);
-        FindDescriptionCommand findSecondCommand = new FindDescriptionCommand(secondPredicate);
+        FindNameCommand findFirstCommand = new FindNameCommand(firstPredicate);
+        FindNameCommand findSecondCommand = new FindNameCommand(secondPredicate);
 
         // same object -> returns true
         assertTrue(findFirstCommand.equals(findFirstCommand));
 
         // same values -> returns true
-        FindDescriptionCommand findFirstCommandCopy = new FindDescriptionCommand(firstPredicate);
+        FindNameCommand findFirstCommandCopy = new FindNameCommand(firstPredicate);
         assertTrue(findFirstCommand.equals(findFirstCommandCopy));
 
         // different types -> returns false
@@ -50,33 +49,25 @@ public class FindDescriptionCommandTest extends FindCommand<FindDescriptionComma
     }
 
     @Test
-    public void equals_notCaseSensitive() throws CommandException {
-        String expectedMessage = String.format(MESSAGE_TASKS_LISTED_OVERVIEW, 1);
-        FindDescriptionCommand command = prepareCommand("op1");
-        assertCommandSuccess(command, expectedMessage, Arrays.asList(PROJECT));
-    }
-
-    @Test
     public void execute_zeroKeywords_noTaskFound() throws CommandException {
         String expectedMessage = String.format(MESSAGE_TASKS_LISTED_OVERVIEW, 0);
-        FindDescriptionCommand command = prepareCommand(" ");
+        FindNameCommand command = prepareCommand(" ");
         assertCommandSuccess(command, expectedMessage, Collections.emptyList());
     }
 
     @Test
     public void execute_multipleKeywords_multipleTasksFound() throws CommandException {
         String expectedMessage = String.format(MESSAGE_TASKS_LISTED_OVERVIEW, 3);
-        FindDescriptionCommand command = prepareCommand("coffee OP1 midterms");
+        FindNameCommand command = prepareCommand("Prepare breakfast Project Revision");
         assertCommandSuccess(command, expectedMessage, Arrays.asList(PREPAREBREAKFAST, PROJECT, REVISION));
     }
 
     /**
-     * Parses {@code userInput} into a {@code FindDescriptionCommand}.
+     * Parses {@code userInput} into a {@code FindNameCommand}.
      */
-    private FindDescriptionCommand prepareCommand(String userInput) {
-        FindDescriptionCommand command =
-            new FindDescriptionCommand(new DescriptionContainsKeywordsPredicate(Arrays.asList(userInput.split
-            ("\\s+"))));
+    private FindNameCommand prepareCommand(String userInput) {
+        FindNameCommand command =
+                new FindNameCommand(new NameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+"))));
         command.setData(model, new CommandHistory(), new UndoRedoStack());
         return command;
     }
